+++ src/main/java/org/mashupmedia/service/MusicManager.java
@@ -38,6 +38,6 @@ public interface MusicManager {
 
 	public void saveArtist(Artist artist);
 
-	public int getTotalSongsFromLibrary(long libraryId);
+	public long getTotalSongsFromLibrary(long libraryId);
 	
 }
